Removes the discontinued items check in weapon parsing.
While this check is ideal to ensure we're not including items that are now discontinued, it adds a ton of latency since each call to page.categories() results in a new HTTP call. There might be a more efficient way to do this, but for now, it's not needed if we deal with exceptions (like when the CombatStyles box is missing).

This also includes some minor fixes to the proto comments at the top of the generated textprotos.